requirements.txt
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
Werkzeug==0.15.5
numpy>=1.9.2
streamlit==1.1.0
tensorflow-cpu==2.7.0opencv-python-headless==4.5.5.62
